+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to fwtv module will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0)
 
+## Unreleased
+
+### Fixed
+
+- fixed a bug where the tolerance was not applied correctly when checking the attendances
+
 ## [2.4.0] - 2025-06-06
 
 ### Added